style for Clothes
