--- src/components/Clothes/ShopFilter/Brands.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Brands = () => (
-  <div>
-    <h2>Brands:</h2>
-    <input type="checkbox"/>Armani
-    <input type="checkbox"/>Versace
-    <input type="checkbox"/>Dolce & Gabbana
-    <input type="checkbox"/>Hugo Boss
-    <input type="checkbox"/>Prada
-  </div>
-);
-
-export default Brands;
