+++ src/components/Clothes/ShopFilter/Brands/Brands.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import './Brands.css';
+
+const Brands = () => (
+  <div className="brands-box">
+    <h2>Brands:</h2>
+	<div className="input-holder"><input type="checkbox"/>Armani</div>
+	<div className="input-holder"><input type="checkbox"/>Versace</div>
+	<div className="input-holder"><input type="checkbox"/>Dolce & Gabbana</div>
+	<div className="input-holder"><input type="checkbox"/>Hugo Boss</div>
+	<div className="input-holder"><input type="checkbox"/>Prada</div>
+  </div>
+);
+
+export default Brands;
