--- src/components/Clothes/ShopFilter/Discaunt.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Discaunt = () => (
-  <div>
-    <h2>Location:</h2>
-    <input type="checkbox"/>5%
-    <input type="checkbox"/>10%
-    <input type="checkbox"/>15%
-    <input type="checkbox"/>20%
-    <input type="checkbox"/>25%
-  </div>
-);
-
-export default Discaunt;
