+++ src/components/Clothes/ShopFilter/Discaunt/Discaunt.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import './Discaunt.css';
+
+const Discaunt = () => (
+    <div className="discaunt-box">
+    <h2>Discaunt:</h2>
+	<div className="input-holder"><input type="checkbox"/>5%</div>
+	<div className="input-holder"><input type="checkbox"/>10%</div>
+	<div className="input-holder"><input type="checkbox"/>15%</div>
+	<div className="input-holder"><input type="checkbox"/>20%</div>
+	<div className="input-holder"><input type="checkbox"/>25%</div>
+  </div>
+);
+
+export default Discaunt;
