--- src/components/Clothes/ShopFilter/Location.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Location = () => (
-  <div>
-    <h2>Location:</h2>
-    <input type="checkbox"/>Sofia
-    <input type="checkbox"/>Plovdiv
-    <input type="checkbox"/>Varna
-    <input type="checkbox"/>Burgas
-    <input type="checkbox"/>Stara Zagora
-  </div>
-);
-
-export default Location;
